[libcolorcorrect] Sync night colour default values
The default value for active in the kcfg file inside kwin for colour
correction is false.

As we have a bridge that mirrors the config contents over DBus and
defaults hardcoded in a second format this has got out of sync.

This both leads to the default button being enabled/disabled incorrectly
on a fresh system and not working properly.

More important as geo-location is on by default this leads to http
requests being made without the user's explicit consent. A major bug.

BUG: 442253