add script + update timezone tables
the new script uses tzdata.zi data to construct the tables of timezone
cities and continents. this way humans don't need to maintain the table.

since timezone city names aren't necessarily correct names there's an
additional mapping table inside the script specifically to map from TZ
city names to the english name we want to use

this still needs running manually every once in a while. we need this
source file up-to-date in the repository so the names can be translated