[Klipper] Handle incoming UTF-8 mime types
QMimeData abstracts most mime type handling, but not text/plain vs
text/plain;charset=utf-8.

This is currently being done in QWayland inside qwaylanddataoffer. Given
this code is effectively a clone, we need to have the same check here.

Interestingly mimeData->text() worked before but not hasText(), which is
why it was unnoticed in the test pasteClient tool.

This fixes klipper and GTK apps.

BUG: 424694