applets/clipboard: Use the phrase "QR code" in the QR code action
This action generates a QR code, but refers to it as a barcode in the UI
strings. In English at least, this is confusing; barcodes and QR codes
are different things, rather than one being a subset of the other. So
let's change the base English text to call the generated image a QR
code rather than a barcode. Any other language where the translation
for "barcode" is appropriate can substitute that in the translated text.