Revert "xembed: adapt to changes in KNotifications"
The file org.kde.StatusNotifierItem was syncronised with up-to-date
master, despite being previous edited to be only supported actions.

Whilst it seems harmless the line:
+    <property name="Menu" type="o" access="read"/>
means we now advertise this property.

Marshalling an empty path gives a warning.

Plasma-workspace "correctly" thinks we have a supported DBus menu and
therefore don't want to receive context menu events.

This reverts commit e782a1248d08adbaa9e3b4ef490c24765c6519c4.
This reverts commit b9de2c8546874efcc993913bd55d0ef252b866a8.

The XML file gains a comment at the top so this doesn't happen again.

BUG: 442758