klipper: Fix reading from plasma on wayland
When copying something inside the same process that klipper is in,
reading the data through wayland will not work since we have to do
so in a blocking way due to how klipper is structured.
But we can check if we currently own the clipboard and just return that.


(cherry picked from commit 01641a2eb6bde1e0e453820e6ead5102d148a2fe)