applets/devicenotifier: Remove applet list item animations
This was always the indended appearance, but it was being overridden by
a bug in the ExpandableListItem itself. Let's remove the animations here
so that the bug in question can be fixed while preserving the intended
visual effect.

(cherry picked from commit 7b0beaae5841eb90a15fbe17292ff8bc7e04148a)