applets/notifications: Embellish header for critical notifications
We get intermittent user complaints and bug reports about notifications
not being visible enough, especially on large or cluttered screens.

This commit attempts to remedy the situation by tinting the header for
on the left side for critical notifications using the color scheme's
"warning" color. Only critical notifications receive this treatment
since they're the only ones that the user really does need to see.

BUG: 420541
FIXED-IN: 5.24