sddm-theme: fix missing password field on "Other" page
On this page there is no concept of the user being passwordless because
no username has been specified by the user yet; we cannot know ahead of
time. As a result, userList.currentItem.needsPassword is blank so the
password field gets hidden.

We should just assume that in username/password entry mode, all users
have a password set. For any users where this is not the case, they can
just leave the password blank, and it's not any worse than it was in
Plasma 5.22.

BUG: 443737
FIXED-IN: 5.23.1