Port away from deprecated AbstarctRunner::actions method
This was deprecated, because having string based key-actions values
does not fit the usecases we have.

Consequently we should check if the runner has any QAction children, this
works if the runner created the QActions manually or if the actions
are registered in the abstract runner.

Also we can drop the ugly hack.