Fix context menu actions for KCMs
Needs https://invent.kde.org/plasma/systemsettings/commit/c815c7a9bc658b7d4778c3fc4ebfff9756ebb199

Currently the logic is very fragile and depends on the data being an url.
However the data of the match is an implementation detail and is in case of the KCM matches
a KPluginMetaData instance.

To avoid this issue the first url of the QueryMatch::urls() return value is checked.
This data is intended for mime-type/scheme related operations.

In the future we need an extra url for systemsettings & should consistently use the
urls of the query match. I will work on this before continuing to port the
plasma KCMs from KServiceTypeTrader.

BUG: 442522