kcm/users: Move "Name" field into the FormLayout
This allows it to be aligned with other form field, which is not
possible at the moment since it exists outside of the FormLayout in a
Row:ayout with the avatar. Fixing this would require adding API to
Kirigami.FormLayout to let its right side align to elements not in
another FormLayout, or port away from FormLayout and use a big
GridLayout instead. Both are messy. Instead, let's put the Name field
in the form layout and center the avatar above it.