Disconnect watcher for xdgActivationTokenArrived
XdgActivationTokens are loaded async. To acheive a job like pattern we
were comparing serials in our slot when one was created.

On wayland this isn't ideal because we're slowly building up an
increasibly long list of lambdas being run, but they'll all no-op.

On X11 because lastInputSerial is always 0 we run the slot multiple
times for every system tray item each time an item is clicked or an
application is launched.

For xembedsniproxy this is especially problematic problematic as we will
trigger synthesised input events in our slot.

BUG: 444385