baloosearchruner: Emit DBus error when baloo is disabled
Otherwise we exit and KRunner runner into a long timeout.

By making this a dedicated DBus error KRunner can also know, that this
runner can not be started. Consequently making new attempts to query
this runner can be prevented.

Because the runner is DBus-Activated, we can not exit the process, otherwise other
processes (like the search in the app launcher) will have the same issue with the timeout.
BUG: 445342