Port kcminit loading away from KServiceTypeTrader
As we have discussed, this mechanism is considered Plasma-internal, meaning
that we don't have to keep the KServiceTypeTrader stuff around.

To avoid duplicating the KCMs or having to split the kcminit stuff in a separate
binary, symlinks are used to query the available plugins. By having a unified "kcminit"
symbol for the function which will get called, we can drop the need for the
custom property.

Task: https://phabricator.kde.org/T14335

Prequisite for using only embedded json metadata for the KCMs, see https://phabricator.kde.org/T14517