[applets/batterymonitor] Refactor data flow in PowerProfileItem
Fixes the following error:

  plasmoids/org.kde.plasma.battery/contents/ui/PowerProfileItem.qml:137:
  TypeError: Cannot read property 'label' of undefined

Profile data was moved to the root component where it makes most sense.
Other chunks of code were aggregated into read-only properties nearby.
It greatly simplified the rest of code in the file.

Took a while to figure out actual structure of the data that comes from
plasma data engine, so now it's somehow documented for future reference.