[applets/batterymonitor] Replace _s with proper Units spacing values
This is still borken as hell, because x/y were not supposed to work
inside Row/Column anyway.

PlasmaCore.Units.largeSpacing is inconsistent with Kirigami, and
actually equals to gridUnit, and 4 times as big as smallSpacing. So,
the proper replacement for a `largeSpacing / 2` (which also properly
takes care of rounding to an integer) is `smallSpacing * 2`. And the
largeSpacing itself shall be replaced with gridUnit to avoid confusion.

GIT_SILENT