[applets/batterymonitor] BatteryItem: Port top-level Column to Layout
This is a direct translation of anchors to QtQuick.Layouts, except
gridUnit in leftMargin is replaced with something it was really meant
to be: the value of a margin on the right side of the batteryIcon,
which happens to be the RowLayout's spacing.

But once we yeet the ToolTipArea (or even before that), those
BatteryDetails and InhibitionHint components should be placed inside
the infoRow's ColumnLayout.