[applets/batterymonitor] BatteryItem: Unwrap BatteryDetails from Loader
It was necessary to drop the `required` qualifier from inListView
property due to a bug in Qt/QML compiler which somehow considers
`required` properties of inline components for named roles resolution,
but only when they are also directly instantiated somewhere in the file
as opposed to deferred instantiation via `Loader::sourceComponent`.

ToolTipArea::active is now bound to just false, which was the case ever
since df2469e7d238ac070f6d87f2c403ce41470dccda. It's going to be removed
completely soon.