[applets/batterymonitor] BatteryItem: Inline BatteryDetails component
Moved aforementioned component's source code directly to the only place
of its instantiation left.

Boolean property `inListView` which was responsible exclusively for the
font size selection, got totally abolished and replaced with font
bindings themselves.

Other boolean properties of the grid were marked as `readonly` which
they should've been from the beginning; and {Left,Right}Labels' binding
to them had to be qualified with property access by id, which --
again -- was just an oversight, and worked only because it was a
top-level(even though an 'inline') component.

Renamed GridLayout's `id` to simply "details" for better readability.