[applets/batterymonitor] PopupDialog: Move battery list inside ColumnLayout
No visually observable changes, again. But it replaces most of the
anchors tweaking with a simple QtQuick.Layout interface.

You might've noticed how

    topMargin: PlasmaCore.Units.gridUnit

got replaced with a different-looking

    Layout.topMargin: PlasmaCore.Units.smallSpacing * 2

but that's all according to the plan, since this ScrollView is now
inside a ColumnLayout which has its own spacing that adds up.

Also, settingsColumn id is removed because it became unused.