[applets/batterymonitor] PowerProfileItem: Refactor Column into Layout
QtQuick.Layouts are easier to manage, and more flexible.

In case of fillWidth+preferredWidth, preferredWidth defines a ratio
rather than an absolute value, so it was replaced with a singleton `1`
instead of a magic number `50`. Since size of this applet propagates
top-down only, it should not affect any calculations, especially
because `50` is too narrow to fit any reasonable text label, except
maybe Chinese.