dataengines/mpris2: Fix empty player condition
m_proxies does not always contain all players, and a player is added to
m_proxies only when it has a valid proxyPid. But the container is
ensured to be added to one of m_playing, m_paused and m_stopped.

So we need to ensure m_playing, m_paused and m_stopped are all empty
before emitting `playerListEmptied()`.

BUG: 447169
FIXED-IN: 5.24