[libtaskmanager] Don't ignore services with NoDisplay=true when searching for appId
For KCMs we have .desktop files in <share>/applications but with NoDisplay=true.

We still want an appId of kcm_foo to match to the service in <share>/applications/kcm_foo.desktop.

The current code however explicitly excludes those, so stop doing that.

This also allows to simplify the code from KServiceTypeTrader to KService::serviceByDesktopName

This fixes the LauncherUrl for the autostart KCM from file:kcmshell to applications:kcm_autostart.desktop