wallpapers: Make "Set as Wallpaper" only apply to current activity
Users may complain after setting a wallpaper to all desktops if they are
using activities, so it's recommended to make the action only apply to the
current activity.

To list all activities in the context menu, it requires a new action in
kf5/kfileitemaction, which seems not possible in desktop file.

CCBUG: 358038