[applets/digitalclock] Allow to launch Kalendar instead of KOrganizer
Currently this is hardcoded to KOrganizer, but Kalendar is a nice app and we should be able to launch that too.

Check for the default handler for calendar files and take that if it is KOrganizer or Kalendar. The check is to avoid launching an app that can technically open ics files but isn't a calendar app, e.g. Kate.

We can extend the logic with other known calendar apps