[libtaskmanager] Consider _GTK_APPLICATION_ID for matching desktop files
Despite our best effors we currently do not match Gedit windows to their desktop file. It manages to evade all of our heuristics.

There is a reliable way to map it and other GTK apps though, the _GTK_APPLICATION_ID window property, which is more or less the same as our desktopFileName property