SystemTray: only make opaque if panel is explicitly specified as opaque
The intention with adaptive opacity was to have increased opacity that didn't look
weird when a window was maximised, with a brightly coloured small space at a screen's
edge against usually plain white or blackish-grey content. In the case of the system tray on top
of that content, this issue doesn't apply, so clashing colours isn't a concern. We can have
the aesthetic of translucency in this case, if the user is in adaptive mode. For the case of
explicitly specified opaque mode, the user probably wants it always opaque (see bug 439025),
so we only become opaque if the user explicitly asks for it, otherwise it's always translucent.