applets/batterymonitor: Improve for desktop systems
As it was so far, it looked like the system was malfunctioning because
we were showing a battery with a big fat red mark. If there's no battery
we can show the default icon.

As for the name, we only say "battery and brightness" when there's
both of them. Otherwise we show either title.

If neither batteries nor brightness are available, we mark the applet as
passive so people can at least still access the power management blocking
checkbox which lives here.

BUG: 415073
FIXED-IN: 5.24