Remove DBus service start condition for baloo runner
Otherwise we do not start the process, which will cause a timeout in KRunner.
The goal of the condition was to prevent the process from starting up
unnecessarily, but having a timeout in KRunner in that case is far from ideal.

With https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/796 the runner
will get disabled when one disables baloo indexing.
Consequently the runner will not get called from DBus.