Use shared contexts for plasmashell
Shared contexts allow us to share textures and vertex buffers between
contexts (windows) this allows us to reduce memory in popups and alike.

To fully make use of this requires some follow up work in
ManagedTextureNode for FrameSVG, but we should sitll get some OOTB.

If shared contexts are not supported this fails gracefully - though we
are relying on shared contexts within kwin for a release without issue.

It also should fix WebGL inside WebViews.

BUG: 448299