[applets/appmenu] Set MenuDelegate's textFormat explicitly to StyledText
Generally, this property should be set on every Text component that we
write. It avoids ambiguity and extra potentially expensive parsing step
with heuristics. Here, we know for sure that the label is in rich text
format.

Text.StyledText format is computationally cheaper than the Text.RichText,
while still getting its job done: displaying the underlined letters when
Alt is key pressed.