applets/devicenotifier: Make "removing" message more instructional
When you unmount a device using the UI, the device shows a message
saying, "Removing..." This message remains visible as long as the device
is still processing file move or copy operations due to the use of async
file IO. The device is not yet safe to unplug as long as the message is
visible, but this is not obvious from the UI. As a result, it is
possible for people to think it's stuck and unplug it anyway, causing
data loss.

This commit attempts to address that by showing a different message when
the unmount takes more than one second, which tells the user what is
actually happening and explicitly asks them not to unplug it yet.