killrunner: Do not define plugin id with a space
This way we no longer crash in kcmutils, because a space is an invalid character for the
DBus registration. While it is fixed in kcmutils 397763b0189f8a61c340e30959391f169ecfa0d, this commit
has not landed in the latest frameworks release.

The only issue is that the plugin would be re-enabled in case it was disabled by the user, because the plugin id is different.
However this runner is not aggressive in the amount of matches it displays, I do not consider this a big issue.

The new plugin Id gets derived from the filename.

BUG: 448156