applets/systemtray: fix glitch for some applets with headers
PlasmoidPopupsContainer.qml has special code to handle merging headers
from applets if those applets use a PlasmaComponents3.Page instance as
their FullRepresentation. However this breaks for any applets that use
PlasmaExtras.Representation, which is a subclass of it that we have been
porting applets to over time, because it offers more features in a
single location. This caused the detection to break, so such applets
woudl have their headers improperly merged.

Explicitly detecting the use of PlasmaExtras.Representation as well
fixes it.

BUG: 438178
FIXED-IN: 5.24