servicesrunner: Fix some apps not showing up when typing less than 3 characters
We want to check the user-visible name, not the name of the file on disk.

In case we type "di", results like "Discord" or "Discover" are expected. With the old
logic however, we would reject discover, because the desktop file is called "org.kde.discover".

BUG: 443866


(cherry picked from commit ac22a4f170d08552a6e7b00f77cd23dc244010fd)