Make SessionManagementScreen a FocusScope
SessionManagementScreen needs to be a FocusScope, so that when it gets focused
directly (through the StackView), it directs it to the proper control inside
instead of stealing it from them.

This fixes initial focus of the controls in the SDDM theme.

BUG: 447817


(cherry picked from commit 10b31ead111ab623879e7796f1434c15097b1f29)