Fix initial focus in LockScreenUi
SessionManagementScreen is a FocusScope and needs focus explicitly assigned
to it. If inside a QQC2 StackView this is taken care of, but for some reason
in a QQC1 StackView this isn't the case for the initial item. Do it explicitly.


(cherry picked from commit a66dcbdda752aaeda2b6163749ae5eb166180259)