[systemd] Fix logout issues
It is important that we end plasma-core.target plasma-workspace.target
and graphical-session.target when we log out. graphical-session in turn
closes all other apps and daemons and prevents restarts.
Handling it in systemd gives the least racey option.

graphical-session used to be hanlded in the plasma-ksmserver service
file, but this was removed in 5.24 in a refactor where it caused issues
for Xwayland restarting.

This patch reintroduces the sentinel programs to close down our target,
which in turn take down the other targets. It also removes a complicated
use of templates that in hindsight didn't really solve anything.

BUG: 448492

Testing done:
- logout of wayland shows all targets as inactive, logging in
repeatedly works smoothly
 - logout of X11 behaves the same.
- calling systemctl --user stop plasma-kwin_wayland.service closes the
session
 - killing ksmserver on wayland does not close the session
 - killing ksmserver on X does

(cherry picked from commit 39ecc45e5bdf8cf167b7b128e2c3473069bd5119)