Klipper: Make the "General" configuration options clearer
Combine the "Text selection only", "Ignore selection" and "Ignore images"
options into separate sets of radio buttons for text and images, to make
their functions explicit.

Remove the "Prevent empty clipboard" option (but leave it as a hidden
config file setting), it was difficult to explain and it is unclear if it
ever had a use case.

Convert the page to a QFormLayout, and add explanatory text to existing
controls where appropriate.

GUI:
I18N: