De-duplicate 'Uninstall or manage addons' option in the context menu of Kickoff
If we have multiple sources for the application, we end up with multiple entries.
This is not useful in any scenario, because the generated appstream URL
is exactly the same.

Consequently we should return from the function when we created the first action.

BUG: 448564


(cherry picked from commit 7024a32824821f7ae76d811fcba08cc16b117aa0)