applets/battery: adjust show/hide conditions for charge limits
When we are using a charge threshold, the kernel may stop charging
within a percentage point of the actual threshold and this is
considered correct behavior, so we have to handle that. See
https://bugzilla.kernel.org/show_bug.cgi?id=215531.

Also, Upower may give us a status of "Not charging" rather than
"Fully charged", so we need to account for that as well. See
https://gitlab.freedesktop.org/upower/upower/-/issues/142.

BUG: 435931


(cherry picked from commit 49013bf4a5fd570fe05910f2d145a926410fc140)