Desktop: Still show services with nodisplay=true set
When adding KCMs to the desktop, we add the entry with NoDisplay=true so that it does
not show up in a search. However before this commit, that would cause the entry to
not show up on the desktop as well.

Because of the filtering in the services runner, we already make sure that only relevant
entries can get added to the desktop in the first place.

BUG: 449243
FIXED-IN: 5.24.1