List "unknown" screens as well
There is also the possibility that the screen mapping by screenpool gets
lost or partially lost ( like removal of plasmashellrc but not
appletsrc)
In this case if it was a multimonitor setup there will be containments
assigned to a screen number which doesn't have a known connector.
List them as well, marked as "Unknown 1" "Unknown 2" etc.
This is an edge case of the edge case but it well possible it happens so
it should be managed as well

![Screenshot_20220125_143614](/uploads/6fc1d1cde79543dca634acc5fd350ba4/Screenshot_20220125_143614.png)