applets/systemtray: fix SNI context menu usage without libappindicator
StatusNotifierItem.qml does not explicitly handle the case of
right-clicking, which causes right-clicks to be ignored unless you also
happen to have the libappindicator package installed, which injects a
handler for this itself. But this should work even if you don't have
that package installed; this commit fixes it.

BUG: 449870
FIXED-IN: 5.24.1