kcms/colors: Optionally apply accent color to titlebars and fix foreground color
This commit gives the accent-overridden pieces of the color scheme overridden
foregrounds as well, preventing readability issues with accents and certain color
schemes when paired up. Additionally, this MR will also prevent readability issues
with lighter colors as the text color is determined between being black or white
by the color's luminosity, aka brightness.

In addition, color schemes can be marked as applying the accent color to active
and/or inactive titlebars. This is exposed to theme creators and can be changed by
users as well.