applets/batterymonitor: Communicate that charging limits are approximate
Kernel developers have said that charge limits are approximate, not
exact[1]. As a result, if for example you have a charge limit of 80%,
your battery might actually stop charging anywhere between 78-82%. But
if we only list the charge limit itself, it looks to users like a KDE
bug when the battery stops charging at any number other than that limit.

To alleviate that, this commit adjusts the text to indicate that the
charge limit is appropximate.

[1] https://bugzilla.kernel.org/show_bug.cgi?id=215531#c3