SDDM theme: stop eliding people's names so aggressively
When more than one user is shown on the login screen, username labels
are limited to the width of the avatar so they don't overlap one
another. However they are not allowed to become multi-line strings, so
they can get elided if they are not somewhat short. This was worsened
recently by a change in Plasma 5.24 to increase the size of the
username labels, which shortened the amount of text that can be seen
without eliding.

This commit fixes that problem by allowing constrained username labels
to become multi-line strings with up to 3 lines, and adjusting the
surrounding layout to visually accommodate this.

BUG: 450673
FIXED-IN: 5.24.3
(cherry picked from commit a630dfb6b6bbe6edc4a0246f309ca6632f763711)