Implement session management file reading without QTextStream
This is needed due to QTextStream::setCodec being gone without replacement
in Qt6. Instead just read the one line we need here via the QFile API. The
input to this file is a raw const char*, so we can read this without
applying any text codec. There's a subtle difference to keep in mind
though, QFile::readLine includes the line break, QTextStream does not.