[Notifications] Move drag handling into a dedicated singleton
`Globals` (popup handling) is a singleton and as such has no access
to context properties like `plasmoid`. To remain "API-compatible"
there is a `plasmoid` property in `Globals.qml`.

However, this broke with the port to a singleton type and is
admittedly bad layering.

Move drag handling into a proper singleton type to avoid
cross-referencing `plasmoid` from ten layers deep down the code.