[Notifications] Provide basic job information in tooltip
Especially for longer downloads where a user is more likely to minimize
the job popup, it is useful to be able to see basic job information
(title, i.e. "Downloads", percentage) without having to open the
notification history.

Moreover, it can be hard to judge completion percentage from the tiny circle.