applets/kicker: Skip creating KService for non-desktop files or folders
When creating a new KService, it will try to read the file as a desktop
file, but if the file is a large non-desktop file, it will hang for a
long time.

Verify the file is a desktop file, and if not, skip creating KService
and run the match directly.

BUG: 442970
FIXED-IN: 5.24.4